refractored app.py, moved helper functions to helper.py
